+++ src_kit/gg_kit.mli
@@ -11,8 +11,9 @@
     is not entirely satisfactory, read the docs. *)
 
 [@@@alert "-unstable"]
-module Pgon2 = Gg__pgon2
 module Color_scheme = Gg__color_scheme
+module Pgon2 = Gg__pgon2
+module Field2 = Gg__field2
 [@@@alert "+unstable"]
 
 (*---------------------------------------------------------------------------